Rename "dirty hands" to "storage tools" && "folder-list" to "ls"
