+++ cmd/sqlserver/sqlserver.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/wal-g/wal-g/cmd/dh"
+	"github.com/wal-g/wal-g/cmd/st"
 
 	"github.com/spf13/cobra"
 	"github.com/wal-g/wal-g/internal"
@@ -39,6 +39,6 @@ func init() {
 	cmd.PersistentFlags().StringVar(&internal.CfgFile, "config", "", "config file (default is $HOME/.walg.json)")
 	cmd.InitDefaultVersionFlag()
 
-	// Dirty hands tool
-	cmd.AddCommand(dh.DirtyHandsCmd)
+	// Storage tools
+	cmd.AddCommand(st.StorageToolsCmd)
 }
